Completed reorganization
